Update the roadmap.
Also, stop guesstimating dates on these things. MoarVM's reason to
exist is Perl 6, and the foucs needs to be on doing what we need for
the Christmas 2015 release. The things on here are all good to have,
but given the almost total overlap between the MoarVM and Perl 6
teams, it's best to let Perl 6's needs drive the priorities.